// frontend/src/api/generated/models/atividadeDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import type { AtividadeDtoFormatoExigido } from './atividadeDtoFormatoExigido';

export interface AtividadeDto {
  id: string;
  materiaId: string;
  titulo: string;
  /** @nullable */
  descricao: string | null;
  formatoExigido: AtividadeDtoFormatoExigido;
  /** @nullable */
  arquivoUrl: string | null;
  /**
     * null só em atividades antigas, criadas antes do prazo ser obrigatório
     * @nullable
     */
  prazo: string | null;
  criadaEm: string;
}